Add format-specifier warnings to SkDebugf.
This CL fixes up many existing format-specifier violations.
Note that GCC has a warning for formatting nothing, so existing calls to
`SkDebugf("")` have been removed, or replaced with `SkDebugf("%s", "")`.
These were apparently meant to be used as a place to set a breakpoint.

+++ src/gpu/vk/GrVkFramebuffer.h
@@ -13,6 +13,8 @@
 #include "src/gpu/vk/GrVkManagedResource.h"
 #include "src/gpu/vk/GrVkResourceProvider.h"
 
+#include <cinttypes>
+
 class GrVkAttachment;
 class GrVkGpu;
 class GrVkImageView;
@@ -57,7 +59,8 @@ public:
 
 #ifdef SK_TRACE_MANAGED_RESOURCES
     void dumpInfo() const override {
-        SkDebugf("GrVkFramebuffer: %d (%d refs)\n", fFramebuffer, this->getRefCnt());
+        SkDebugf("GrVkFramebuffer: %" PRIdPTR " (%d refs)\n",
+                 (intptr_t)fFramebuffer, this->getRefCnt());
     }
 #endif
